Reorganize Prayer Group Types
