+++ components/prayer-group/create-prayer-request/create-prayer-request-types.ts
@@ -1,9 +1,3 @@
-import { SortConfig } from "../../../types/api-response-types";
-import {
-  RawPrayerGroupSummary,
-  RawPrayerGroupUserSummary,
-} from "../../../types/prayer-group-types";
-
 export enum TimeToLiveOption {
   OneWeek = 7,
   TwoWeeks = 14,
@@ -27,29 +21,3 @@ export enum CreatePrayerRequestWizardStep {
   RequestBodyStep = "RequestBodyStep",
   ExpirationStep = "ExpirationStep",
 }
-
-export type PrayerRequestFilterCriteria = {
-  prayerGroupIds?: number[];
-  creatorUserIds?: number[];
-  pageIndex?: number;
-  pageSize?: number;
-  bookmarkedByUserId?: number;
-  includeExpiredRequests?: boolean;
-  sortConfig: SortConfig;
-};
-
-export type RawPrayerRequestModel = {
-  id?: number;
-  requestTitle?: string;
-  requestDescription?: string;
-  createdDate?: string;
-  prayerGroup?: RawPrayerGroupSummary;
-  user?: RawPrayerGroupUserSummary;
-  likeCount?: number;
-  commentCount?: number;
-  prayedCount?: number;
-  expirationDate?: string;
-  isUserPrayed?: boolean;
-  isUserLiked?: boolean;
-  isUserCommented?: boolean;
-};
