+++ components/prayer-group/prayer-group-types.ts
@@ -1,0 +1,31 @@
+import { SortConfig } from "../../types/api-response-types";
+import {
+  RawPrayerGroupSummary,
+  RawPrayerGroupUserSummary,
+} from "../../types/prayer-group-types";
+
+export type PrayerRequestFilterCriteria = {
+  prayerGroupIds?: number[];
+  creatorUserIds?: number[];
+  pageIndex?: number;
+  pageSize?: number;
+  bookmarkedByUserId?: number;
+  includeExpiredRequests?: boolean;
+  sortConfig: SortConfig;
+};
+
+export type RawPrayerRequestModel = {
+  id?: number;
+  requestTitle?: string;
+  requestDescription?: string;
+  createdDate?: string;
+  prayerGroup?: RawPrayerGroupSummary;
+  user?: RawPrayerGroupUserSummary;
+  likeCount?: number;
+  commentCount?: number;
+  prayedCount?: number;
+  expirationDate?: string;
+  isUserPrayed?: boolean;
+  isUserLiked?: boolean;
+  isUserCommented?: boolean;
+};
